pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aggrequant"
version = "2.0.0"
description = "Automated aggregate quantification for High Content Screening"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Athena Economides"}
]
keywords = ["image-analysis", "high-content-screening", "segmentation", "deep-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Topic :: Scientific/Engineering :: Image Processing",
]
dependencies = [
    # Image processing
    "numpy>=1.24",
    "scikit-image>=0.21",
    "opencv-python-headless>=4.8",
    "tifffile>=2023.7",

    # Data handling
    "pandas>=2.0",
    "pyarrow>=14.0",

    # Configuration
    "pyyaml>=6.0",
    "pydantic>=2.0",

    # CLI
    "click>=8.1",
    "tqdm>=4.65",

    # Visualization
    "matplotlib>=3.7",
]

[project.optional-dependencies]
# For running the segmentation pipeline
pipeline = [
    "torch>=2.0",
    "torchvision>=0.15",
    "stardist>=0.8",
    "cellpose>=3.0,<4.0",  # v3 API required
]

# For training neural networks
nn-training = [
    "torch>=2.0",
    "torchvision>=0.15",
    "segmentation-models-pytorch>=0.3",
    "albumentations>=1.3",
    "tensorboard>=2.15",
]

# For the web GUI
gui = [
    "dash>=4.0",
    "plotly>=5.0",
]

# All dependencies
all = [
    "aggrequant[pipeline,nn-training,gui]",
]

# Development
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
]

[project.scripts]
aggrequant = "aggrequant.cli:main"
aggrequant-train = "aggrequant.nn.cli:train"
aggrequant-benchmark = "aggrequant.nn.cli:benchmark"
aggrequant-gui = "aggrequant.gui_web.app:main"

[project.urls]
Homepage = "https://github.com/photonics-crete/AggreQuant"
Repository = "https://github.com/photonics-crete/AggreQuant"

[tool.setuptools.packages.find]
include = ["aggrequant*"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
